% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrrrl.R
\name{general}
\alias{general}
\alias{FROM}
\alias{WHERE}
\alias{AND}
\alias{OR}
\alias{GROUP_BY}
\alias{LIMIT}
\alias{DESC}
\alias{ORDER_BY}
\title{General SQL Snippet Functions}
\usage{
FROM(...)

WHERE(..., cond = TRUE)

AND(...)

OR(...)

GROUP_BY(...)

LIMIT(n = 1)

DESC(x)

ORDER_BY(...)
}
\description{
Create SQL snippets to build SQL queries.
}
\section{Functions}{
\itemize{
\item \code{FROM}: Create FROM SQL snippet with optional table aliases

\item \code{WHERE}: Generate WHERE SQL snippet if \code{cond} evaluates to TRUE, with
arguments to WHERE concatenated by \code{\link{AND}}

\item \code{AND}: Concatenate arguments with \code{AND}

\item \code{OR}: Concatenate arguments with \code{OR}

\item \code{GROUP_BY}: Create \code{GROUP BY} SQL snippet, arguments are column names
separated by commas

\item \code{LIMIT}: Create \code{LIMIT} SQL snippet

\item \code{DESC}: Add \code{DESC} or \code{ASC} after column name

\item \code{ORDER_BY}: Create \code{ORDER BY} SQL snippet, arguments are column names
separated by commas
}}

\examples{
FROM('table1', 'table2')
FROM(t1 = 'table1', t2 = 'table2')
WHERE('col1 IS NOT NULL')
WHERE(cond = TRUE, 'col1 = 2', 'col2 >= 10')
WHERE(cond = FALSE, 'col1 = 2', 'col2 >= 10')
AND(eq(id = 3, class = 'text_value'), geq(date = '2017-06-14'))
OR(eq(id = 9, id = 12), leq(id = 5))
GROUP_BY('col1', 'col2', 'col3')
LIMIT(10)
ORDER_BY('col1', 'col2', 'col3')
ORDER_BY(DESC('col1'), 'col2', ASC('col3'))
}
