% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrrl.R
\name{INSERT_INTO_VALUES}
\alias{INSERT_INTO_VALUES}
\title{INSERT INTO tbl VALUES}
\usage{
INSERT_INTO_VALUES(tbl, vals, cols = NULL)
}
\arguments{
\item{tbl}{Table name to insert into}

\item{vals}{Values for insertion, can be a vector or a data.frame}

\item{cols}{Columns to include}
}
\description{
Inserts into a table the values in a vector or data.frame. Subsets the values
by column name.
}
\examples{
INSERT_INTO_VALUES('table', 1:3)
INSERT_INTO_VALUES('table', c('a' = 1, 'b' = 2, 'c' = 3))
INSERT_INTO_VALUES('iris', iris[c(1, 51, 101), ], c("Sepal.Length", "Petal.Length", "Species"))

}
