% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrrrl.R
\name{SELECT}
\alias{SELECT}
\alias{SELECT_}
\alias{SELECT_DISTINCT}
\title{Generate SELECT statement snippets}
\usage{
SELECT(..., .distinct = FALSE)

SELECT_(table_cols, .distinct = FALSE)

SELECT_DISTINCT(...)
}
\arguments{
\item{...}{Columns to select. Single arguments are assumed to be column
names. Single named arguments are renamed columns. Named arguments with a
vector of length greater than 1 or a vector with named entries are assumed
to be table names; entry names are column aliases.}

\item{.distinct}{If true, uses SELECT DISTINCT as keyword}

\item{table_cols}{Named list of tables and columns}
}
\description{
Generate an SQL snippet for SELECT or SELECT DISTINCT
}
\section{Functions}{
\itemize{
\item \code{SELECT_}: Standard eval version of SELECT

\item \code{SELECT_DISTINCT}: Alias for \code{SELECT(..., .distinct = TRUE)}
}}

\examples{
SELECT(letters[1:3])
SELECT(letters[1:3], 't2' = letters[4:6])
SELECT(a = 'apple', b = 'banana', c = 'cherry')
SELECT('t1' = c(a = 'apple', b = 'banana'), c = 'cherry')
SELECT('t1' = c(a = 'apple', b = 'banana'), c = 'cherry', 't2' = c(d = 'dragon_fruit'))

SELECT_(list('t1' = c('a', z = 'b'), 't2' = 'c'))

}
