% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{sql_typeof}
\alias{sql_typeof}
\title{Returns output type of a vector or object relative to SQL}
\usage{
sql_typeof(x)
}
\description{
Returns best-case (or best-guess) option for the SQL type of the input.
Mostly used to determine quoting rules.
}
