% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlformat.R
\name{sqlformat}
\alias{sqlformat}
\alias{sqlsimplify}
\title{Format an SQL Query using sqlformat}
\usage{
sqlformat(sql, header = NULL, md = TRUE,
  sqlformat_options = "-k upper -a")

sqlsimplify(sql, ...)
}
\arguments{
\item{sql}{SQL Query for formatting}

\item{header}{Header added to formatted query as a comment}

\item{md}{Add markdown backticks?}

\item{sqlformat_options}{Options passed to \code{sqlformat}, see \code{sqlformat -h} for more information.}
}
\description{
Uses the python package \href{https://github.com/andialbrecht/sqlparse}{andialbrecht/sqlparse}
to format an SQL query. See \code{sqlformat -h} for more information about formatting options.
}
\section{Functions}{
\itemize{
\item \code{sqlsimplify}: Wrapper around \link{sqlformat} to simplify SQL statements.
}}

