% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrrl.R
\name{UPDATE}
\alias{UPDATE}
\title{Create \code{UPDATE} SQL statement}
\usage{
UPDATE(table, set, ..., .ignore = FALSE, .order = NULL, .limit = NULL)
}
\arguments{
\item{table}{Table name for update (can be named)}

\item{set}{Named vector of column-value pairs, where the vector name is the
column name.}

\item{...}{Conditions passed on to \code{WHERE} clause (optional)}

\item{.ignore}{Add \code{IGNORE} keyword to \code{UPDATE} clause}

\item{.order}{Optional vector of columns passed to \code{\link{ORDER_BY}}}

\item{.limit}{Optional number of rows for \code{LIMIT} condition}
}
\description{
Creates an UPDATE SQL statement for a single table. Currently, this function
cannot handle using expressions in the \code{set} parameter. \code{UPDATE} is only
useful at this time for single table updates when \code{set} contains actual
values.
}
\details{
Improvements are planned for this function to handle multiple table updates
and the use of SQL expressions in the \code{SET} statement.
}
\examples{
UPDATE('iris', c(some_column = 1, some_other_col = "high"), eq(another_col = 2), geq(a_third_col = 10))
UPDATE('t1', c(col1 = 'a'))
UPDATE('t1', c(col1 = 'a', col2 = 42), 'id' \%IN\% 1:5)
UPDATE('t', c(id = 'id + 1'), .order = DESC('id'))

}
