% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrrl.R
\name{UPDATE}
\alias{UPDATE}
\title{UPDATE}
\usage{
UPDATE(tables, col_vals, ..., .ignore = FALSE, .order = NULL,
  .limit = NULL)
}
\arguments{
\item{tables}{Table name(s) for update (can be named)}

\item{...}{Conditions passed on to \code{WHERE} clause (optional)}

\item{.ignore}{Add \code{IGNORE} keyword to \code{UPDATE} clause}

\item{.order}{Optional vector of columns passed to \code{\link{ORDER_BY}}}

\item{.limit}{Optional number of rows for \code{LIMIT} condition}

\item{set}{Named vector of column-value pairs, where the vector name is the
column name.}
}
\description{
Create \code{UPDATE} SQL statement
}
\examples{
UPDATE('iris', c(some_column = 1, some_other_col = "high"), eq(another_col = 2), geq(a_third_col = 10))
UPDATE('t1', c(col1 = 'col1 + 1'))
UPDATE('t1', c(col1 = 'col1 + 1', col2 = 'col1'))
UPDATE('t', c(id = 'id + 1'), .order = DESC('id'))

}
